moving old salzinnes heuristics into own folder
